requirements.txt
gunicorn
gevent
flask_cors
flask_script
peewee
pymysql
requests
cython
python-dotenv
